+++ lib/src/auth/bearer_client.dart
@@ -1,3 +1,0 @@
-export 'bearer_client_io.dart'
-    if (dart.library.io) 'bearer_client_io.dart'
-    if (dart.library.js_interop) 'bearer_client_web.dart';